refactor: consolidate BaseDAO test files into single comprehensive test file
Merged test_base_dao.py, base_test.py, and base_dao_test.py into a single
base_dao_test.py file. This consolidation reduces redundancy and improves
test organization by grouping all BaseDAO-related tests in one location.

The combined file maintains all test coverage including:
- Column operator enum tests
- Find by ID/IDs operations with flexible column support
- CRUD operations (create, read, update, delete)
- List and filtering functionality
- SQLAlchemy error handling
- Edge cases and special scenarios
